fix(test): address CodeRabbit review — test boundaries + diagnostic log
- Remove Japanese exact-keys from ConversationDisplayTextPatchTests
  so tests verify pass-through, not identity dictionary lookup
- Move 3 MessageLogPatch observation-only tests to L1 fixture
- Add diagnostic log to XDidYTranslationPatch.ResolveValidateMethod
