+++ Mods/QudJP/Assemblies/QudJP.Tests/L2/MessageLogPatchTests.cs
@@ -407,36 +407,6 @@ public sealed class MessageLogPatchTests
         }
     }
 
-    [Test]
-    public void Prefix_ObservationOnly_DoesNotTranslateMessage()
-    {
-        SinkObservation.ResetForTests();
-        var message = "You hit the bear.";
-        MessageLogPatch.Prefix(ref message);
-        Assert.That(message, Is.EqualTo("You hit the bear."));
-    }
-
-    [Test]
-    public void Prefix_ObservationOnly_LogsUnclaimed()
-    {
-        SinkObservation.ResetForTests();
-        var message = "You hit the bear.";
-        var originalMessage = message;
-        MessageLogPatch.Prefix(ref message);
-        var hitCount = SinkObservation.GetHitCountForTests(
-            nameof(MessageLogPatch), nameof(MessageLogPatch), SinkObservation.ObservationOnlyDetail, originalMessage, originalMessage);
-        Assert.That(hitCount, Is.GreaterThan(0));
-    }
-
-    [Test]
-    public void Prefix_DirectMarker_StillStripped()
-    {
-        var message = "\u0001すでに翻訳済みテキスト";
-        var result = MessageLogPatch.Prefix(ref message);
-        Assert.That(message, Is.EqualTo("すでに翻訳済みテキスト"));
-        Assert.That(result, Is.True);
-    }
-
     private static string CreateHarmonyId()
     {
         return $"qudjp.tests.{Guid.NewGuid():N}";
