+++ Mods/QudJP/Assemblies/src/Patches/XDidYTranslationPatch.cs
@@ -30,7 +30,12 @@ public static class XDidYTranslationPatch
         if (GameObjectType is null) return null;
         // Validate has two overloads: Validate(ref GameObject) and Validate(GameObject).
         // Use the non-ref overload for simpler invocation.
-        return AccessTools.Method(GameObjectType, "Validate", new[] { GameObjectType });
+        var method = AccessTools.Method(GameObjectType, "Validate", new[] { GameObjectType });
+        if (method is null)
+        {
+            Trace.TraceError("QudJP: XDidYTranslationPatch failed to resolve GameObject.Validate(GameObject). Validation will be skipped.");
+        }
+        return method;
     }
     private static readonly MethodInfo? ConsequentialColorMethod = AccessTools.Method("XRL.Messages.ColorCoding:ConsequentialColor");
